requirements.txt
# Flask Packages
Flask==3.1.2

# Other Packages
llama-index==0.14.14
python-decouple==3.8
